+++ src/main/java/aztech/modern_industrialization/machines/helper/SteamHelper.java
@@ -34,11 +34,11 @@ public final class SteamHelper {
     public static long consumeSteamEu(List<ConfigurableFluidStack> fluidStacks, long maxEu, Simulation simulation) {
         Preconditions.checkArgument(maxEu >= 0, "May not consume < 0 EU.");
 
-        long totalRem = 0;
+        int totalRem = 0;
         for (ConfigurableFluidStack stack : fluidStacks) {
             if (stack.getResource().getFluid() == MIFluids.STEAM.asFluid()) {
-                long amount = stack.getAmount();
-                long rem = Math.min(maxEu, amount);
+                int amount = stack.getAmount();
+                int rem = (int) Math.min(maxEu, amount);
                 if (simulation.isActing()) {
                     stack.decrement(rem);
                 }
